pyproject.toml
[project]
name = "tiny-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "baidusearch>=1.0.3",
    "duckduckgo-search>=8.0.1",
    "firecrawl-py>=2.4.0",
    "googlesearch-python~=1.3.0",
    "lxml[html-clean]>=5.4.0",
    "openai>=1.76.0",
    "playwright==1.50.0",
    "prompt-toolkit>=3.0.51",
    "readabilipy>=0.3.0",
    "readability-lxml>=0.8.1",
    "rich>=14.0.0",
    "tiktoken>=0.9.0",
    "typer[all]>=0.15.2",
]
